empire-bot/requirements.txt
python-telegram-bot[job-queue]==21.10
httpx==0.28.1
docker==7.1.0
python-dotenv==1.0.1
